Convert to use a SQLite Backend
Source: R/as_sqlite.R

Converts objects to use a SQLite backend, depending on the input type:

data.frame: Converts to a DataBackendDplyr. The primary key column can be provided. If this column does not exists, it will be created.

[mlr3::DataBackend]: Creates a new DataBackendDplyr using the data of the provided mlr3::DataBackend.

[mlr3::Task]: Replaces the DataBackend in slot $task with a new backend. Only active columns and rows are considered.

as_sqlite(data, path = NULL, ...)

Arguments

data
(data.frame | mlr3::DataBackend | mlr3::Task)
See description.

path
(NULL | character(1))
Path for the SQLite data base. Defaults to a temporary file in the R session directory, see tempfile().

primary_key
(character(1))
Name of the primary key column.

Value

DataBackendDbplyr.