DataBackend for dplyr/dbplyr

A mlr3::DataBackend using tbl from packages dplyr/dbplyr. Allows to connect a Task to a out-of-memory data base.

Returns an object of class mlr3::DataBackend.

Usage

# Construction
b = DataBackendDplyr$new(tbl, primary_key)

The interface is described in mlr3::DataBackend.

Arguments

tbl [tbl]
See dplyr::tbl() for construction. Also note that all tibbles inherit from tbl.

primary_key [character(1)]:
Name of the column in tbl which represents a unique row identifier (as integer or character).

Examples

# Backend using a in-memory tibble
data = tibble::as.tibble(iris)
data$row_id = 1:150
b = DataBackendDplyr$new(data, primary_key = "row_id")

# Object supports all accessors of DataBackend
print(b)

#> DataBackend <DataBackendDbplyr> (150x6)
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species row_id
#> 1:          5.1         3.5          1.4         0.2  setosa      1
#> 2:          4.9         3.0          1.4         0.2  setosa      2
#> 3:          4.7         3.2          1.3         0.2  setosa      3
#> 4:          4.6         3.1          1.5         0.2  setosa      4
#> 5:          5.0         3.6          1.4         0.2  setosa      5
#> 6:          5.4         3.9          1.7         0.4  setosa      6

b$nrow

#> [1] 150

b$ncol

#> [1] 6

b$colnames

#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"     
#> [6] "row_id"      

b$data(rows = 100:101, cols = "Species")

#>       Species
#> 1: versicolor
#> 2:  virginica

b$distinct("Species")

#> $Species
#> [1] "setosa"     "versicolor" "virginica" 
#> 

# Classification task using this backend
task = mlr3::TaskClassif$new(id = "iris_tibble", backend = b, target = "Species")
print(task)

#> Task 'iris_tibble' of type classif (150 x 5)
#> Target: Species
#> Features: Sepal.Length, Sepal.Width, Petal.Length, Petal.Width
#> 
#> Public: all_classes, backend, cbind, class_n, class_names, clone,
#>   col_info, col_roles, data, feature_names, feature_types, filter,
#>   formula, hash, head, id, levels, measures, ncol, negative, nrow,
#>   overwrite, positive, properties, rbind, row_ids, row_roles, select,
#>   set_col_role, set_row_role, target_names, task_type, truth

task$head()

#>    Species Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1:  setosa          5.1         3.5          1.4         0.2
#> 2:  setosa          4.9         3.0          1.4         0.2
#> 3:  setosa          4.7         3.2          1.3         0.2
#> 4:  setosa          4.6         3.1          1.5         0.2
#> 5:  setosa          5.0         3.6          1.4         0.2
#> 6:  setosa          5.4         3.9          1.7         0.4

# Create a temporary SQLite data base
con = DBI::dbConnect(RSQLite::SQLite(), ":memory:")
dplyr::copy_to(con, data)
tbl = dplyr::tbl(con, "data")

# Define a backend on a subset of the data base
tbl = dplyr::select_at(tbl, setdiff(colnames(tbl), "Sepal.Width")) # do not use column "Sepal.Width"
tbl = dplyr::filter(tbl, row_id %in% 1:120) # Use only first 120 rows
b = DataBackendDplyr$new(tbl, primary_key = "row_id")
print(b)

#> DataBackend <DataBackendDbplyr> (120x5)
#>    Sepal.Length Petal.Length Petal.Width Species row_id
#> 1:          5.1          1.4         0.2  setosa      1
#> 2:          4.9          1.4         0.2  setosa      2
#> 3:          4.7          1.3         0.2  setosa      3
#> 4:          4.6          1.5         0.2  setosa      4
#> 5:          5.0          1.4         0.2  setosa      5
#> 6:          5.4          1.7         0.4  setosa      6

# Note that SQLite does not support factors, column Species has been converted to character
lapply(b$head(), class)

#> $Sepal.Length
#> [1] "numeric"
#> 
#> $Petal.Length
#> [1] "numeric"
#> 
#> $Petal.Width
#> [1] "numeric"
#> 
#> $Species
#> [1] "character"
#> 
#> $row_id
#> [1] "integer"
#> 

# Cleanup
rm(tbl)
DBI::dbDisconnect(con)